requirements.txt
webdriver-manager
selenium
allure-behave
behave
junit2html